# openapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spring Application exposes Api to manage configuration for EC/PSP on the Nodo dei Pagamenti",
    "termsOfService": "https://www.pagopa.gov.it/",
    "title": "PagoPA API configuration",
    "version": "0.2.28"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/apiconfig/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/services": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "default": 50,
            "description": "Number of elements on one page. Default = 50",
            "format": "int32",
            "in": "query",
            "name": "limit",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Page number. Page value starts from 0",
            "format": "int32",
            "in": "query",
            "name": "page",
            "required": true,
            "type": "integer"
          },
          {
            "in": "query",
            "name": "pspcode",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "brokerpspcode",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "channelcode",
            "required": false,
            "type": "string"
          },
          {
            "format": "int64",
            "in": "query",
            "name": "paymentmethodchannel",
            "required": false,
            "type": "integer"
          },
          {
            "enum": [
              "BBT",
              "BP",
              "AD",
              "CP",
              "PO",
              "OBEP",
              "JIF",
              "MYBK",
              "PPAL"
            ],
            "in": "query",
            "name": "paymenttypecode",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "pspflagftamp",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "channelapp",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "onus",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "flagio",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "flowid",
            "required": false,
            "type": "string"
          },
          {
            "format": "double",
            "in": "query",
            "name": "minimumamount",
            "required": false,
            "type": "number"
          },
          {
            "format": "double",
            "in": "query",
            "name": "maximumamount",
            "required": false,
            "type": "number"
          },
          {
            "default": "IT",
            "enum": [
              "IT",
              "EN",
              "FR",
              "DE",
              "SL"
            ],
            "in": "query",
            "name": "languagecode",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "conventionCode",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Services"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Payment Service Providers"
        ],
        "operationId": "getServices",
        "summary": "Get paginated list of services"
      }
    },
...
  },
  "definitions": {
...
    "PageInfo": {
      "properties": {
        "items_found": {
          "description": "Number of items found. (The last page may have fewer elements than required)",
          "format": "int32",
          "type": "integer"
        },
        "limit": {
          "description": "Required number of items per page",
          "format": "int32",
          "type": "integer"
        },
        "page": {
          "description": "Page number",
          "format": "int32",
          "type": "integer"
        },
        "total_pages": {
          "description": "Total number of pages",
          "format": "int32",
          "type": "integer"
        }
      },
      "required": [
        "items_found",
        "limit",
        "page",
        "total_pages"
      ],
      "type": "object"
    },
...
    "ProblemJson": {
      "properties": {
        "detail": {
          "description": "A human readable explanation specific to this occurrence of the problem.",
          "example": "There was an error processing the request",
          "type": "string"
        },
        "status": {
          "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
          "example": 200,
          "format": "int32",
          "maximum": 600,
          "minimum": 100,
          "type": "integer"
        },
        "title": {
          "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "Service": {
      "properties": {
        "tags": {
          "maxLength": 135,
          "minLength": 0,
          "type": "string"
        },
        "abi_code": {
          "maxLength": 5,
          "minLength": 0,
          "type": "string"
        },
        "broker_psp_code": {
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "cart_card": {
          "type": "boolean"
        },
        "channel_app": {
          "type": "boolean"
        },
        "channel_code": {
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "channel_url": {
          "type": "string"
        },
        "convention_code": {
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "fixed_cost": {
          "format": "double",
          "type": "number"
        },
        "flag_io": {
          "type": "boolean"
        },
        "flow_id": {
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "language_code": {
          "enum": [
            "IT",
            "EN",
            "FR",
            "DE",
            "SL"
          ],
          "maxLength": 2,
          "minLength": 0,
          "type": "string"
        },
        "logo_psp": {
          "items": {
            "format": "byte",
            "type": "string"
          },
          "type": "array"
        },
        "logo_service": {
          "items": {
            "format": "byte",
            "type": "string"
          },
          "type": "array"
        },
        "maximum_amount": {
          "format": "double",
          "type": "number"
        },
        "minimum_amount": {
          "format": "double",
          "type": "number"
        },
        "mybank_code": {
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "on_us": {
          "type": "boolean"
        },
        "payment_method_channel": {
          "format": "int64",
          "type": "integer"
        },
        "payment_type_code": {
          "enum": [
            "BBT",
            "BP",
            "AD",
            "CP",
            "PO",
            "OBEP",
            "JIF",
            "MYBK",
            "PPAL"
          ],
          "type": "string"
        },
        "psp_business_name": {
          "type": "string"
        },
        "psp_code": {
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "psp_flag_stamp": {
          "type": "boolean"
        },
        "service_availability": {
          "maxLength": 511,
          "minLength": 0,
          "type": "string"
        },
        "service_description": {
          "maxLength": 511,
          "minLength": 0,
          "type": "string"
        },
        "service_name": {
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "timestamp_insertion": {
          "format": "date-time",
          "type": "string"
        },
        "validity_date": {
          "format": "date-time",
          "type": "string"
        }
      },
      "type": "object"
    },
    "Services": {
      "properties": {
        "page_info": {
          "$ref": "#/definitions/PageInfo"
        },
        "services": {
          "items": {
            "$ref": "#/definitions/Service"
          },
          "type": "array"
        }
      },
      "required": [
        "page_info",
        "services"
      ],
      "type": "object"
    },
...
  },
  "securityDefinitions": {
    "ApiKey": {
      "description": "The API key to access this function app.",
      "in": "header",
      "name": "Ocp-Apim-Subscription-Key",
      "type": "apiKey"
    },
    "Authorization": {
      "bearerFormat": "JWT",
      "description": "JWT token get after Azure Login",
      "scheme": "Bearer",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
...
}